Implement morton type independent allocation
Allocation methods should be independent of the actual data type of the
morton code, i.e. i32 or i64. The code works but needs further testing.

+++ lib/octree_defines.h
@@ -36,7 +36,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #define BLOCK_SIDE 8
 
-typedef unsigned int morton_type; 
+typedef unsigned long long int morton_type; 
 //   typedef long long int morton_type; 
 
 #endif
